+++ Data/Models/Test.cs
@@ -7,13 +7,13 @@ namespace Data.Models
     [Table("Tests")]
     public class Test : ITest
     {
-        public Test(int testId, int testParametersId, int userId, int endpointId, double databaseTestTime, double applicationTestTime, double apiTestTime, DateTime timeStamp) 
+        public Test(long testId, long testParametersId, long userId, long endpointId, double databaseTestTime, double applicationTestTime, double apiTestTime, DateTime timeStamp) 
             : this(testParametersId, userId, endpointId, databaseTestTime, applicationTestTime, apiTestTime, timeStamp)
         {
             TestId = testId;
         }
 
-        public Test(int testParametersId, int userId, int endpointId, double databaseTestTime, double applicationTestTime, double apiTestTime, DateTime timeStamp)
+        public Test(long testParametersId, long userId, long endpointId, double databaseTestTime, double applicationTestTime, double apiTestTime, DateTime timeStamp)
         {
             TestParametersId = testParametersId;
             UserId = userId;
@@ -24,10 +24,10 @@ namespace Data.Models
             TimeStamp = timeStamp;
         }
 
-        public int TestId { get; set; }
-        public int TestParametersId { get; set; }
-        public int UserId { get; set; }
-        public int EndpointId { get; set; }
+        public long TestId { get; set; }
+        public long TestParametersId { get; set; }
+        public long UserId { get; set; }
+        public long EndpointId { get; set; }
         public double DatabaseTestTime { get; set; }
         public double ApplicationTestTime { get; set; }
         public double ApiTestTime { get; set; }
