+++ TestLibrary/BusinessObject/Test.cs
@@ -5,14 +5,14 @@ namespace TestLibrary.BusinessObject
 {
     public class Test : ITest
     {
-        public Test(int testId, int testParametersId, int userId, int endpointId, double databaseTestTime, double applicationTestTime, double apiTestTime, DateTime timeStamp) 
+        public Test(long testId, long testParametersId, long userId, long endpointId, double databaseTestTime, double applicationTestTime, double apiTestTime, DateTime timeStamp) 
             : this(testParametersId, userId, endpointId, databaseTestTime, applicationTestTime, apiTestTime)
         {
             TestId = testId;
             TimeStamp = timeStamp;
         }
 
-        public Test(int testParametersId, int userId, int endpointId, double databaseTestTime, double applicationTestTime, double apiTestTime)
+        public Test(long testParametersId, long userId, long endpointId, double databaseTestTime, double applicationTestTime, double apiTestTime)
         {
             TestParametersId = testParametersId;
             UserId = userId;
@@ -23,10 +23,10 @@ namespace TestLibrary.BusinessObject
             TimeStamp = DateTime.Now;;
         }
 
-        public int TestId { get; }
-        public int TestParametersId { get; }
-        public int UserId { get; }
-        public int EndpointId { get; }
+        public long TestId { get; }
+        public long TestParametersId { get; }
+        public long UserId { get; }
+        public long EndpointId { get; }
         public double DatabaseTestTime { get; }
         public double ApplicationTestTime { get; }
         public double ApiTestTime { get; }
